+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Development
 
+- Change modelling of HVD-categorisation to DCAT-AP.de 3.0:
+  - Instead of using `dct:references`, use `dcatap:hvdCategory`.
+  - Use `dcatap:applicableLegislation <http://data.europa.eu/eli/reg_impl/2023/138/oj>`.
+
 ## [0.3.9](https://github.com/berlinonline/ckanext-dcatde_berlin/releases/tag/0.3.9)
 
 _(2025-05-07)_