+++ CHANGELOG.md
@@ -2,11 +2,17 @@
 
 ## Development
 
+## [0.4.0](https://github.com/berlinonline/ckanext-dcatde_berlin/releases/tag/0.4.0)
+
+_(2025-05-12)_
+
 - Change modelling of HVD-categorisation to DCAT-AP.de 3.0:
   - Instead of using `dct:references`, use `dcatap:hvdCategory`.
   - Use `dcatap:applicableLegislation <http://data.europa.eu/eli/reg_impl/2023/138/oj>`.
 - Deprecate `politicalGeocodingURI`, use `dct:spatial`.
-- Optionally link to https://berlinonline.github.io/lod-berlin-lor/ when using `dct:spatial`
+- Optionally link to https://berlinonline.github.io/lod-berlin-lor/ when using `dct:spatial`.
+- Use `Namespace` definitions from `rdflib.namespace` instead of defining them ourselves, whenever possible.
+- Add `dcat:temporalResolution` statements based on our `temporal_granularity` attribute.
 
 ## [0.3.9](https://github.com/berlinonline/ckanext-dcatde_berlin/releases/tag/0.3.9)